requirements.txt
customtkinter==5.2.2
Pillow>=11.0
pywin32>=311
tkinterdnd2==0.4.3
CTkMessagebox==2.6
